+++ CHANGELOG.md
@@ -14,6 +14,7 @@ Changelog
 
 **Next release**
 
+* Fixed all minor bugs reported through the support info
 * Added info icon to toggle restrictions
 
 [Open issues](https://github.com/M66B/XPrivacy/issues?state=open)