+++ CHANGELOG.md
@@ -15,6 +15,7 @@ Changelog
 **Next release**
 
 * Fixed all minor bugs reported through the support info
+* Fixed some minor bugs and improved some layouts
 * Added info icon to toggle restrictions
 
 [Open issues](https://github.com/M66B/XPrivacy/issues?state=open)